[PromotionBundle] Move logic from PromotionRuleGroupValidator to PromotionRuleValidator
