+++ src/Sylius/Bundle/PromotionBundle/spec/Validator/PromotionRuleValidatorSpec.php
@@ -20,13 +20,18 @@ use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
 use Symfony\Component\Validator\Exception\UnexpectedValueException;
+use Symfony\Component\Validator\Validator\ContextualValidatorInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Validator\Violation\ConstraintViolationBuilderInterface;
 
 final class PromotionRuleValidatorSpec extends ObjectBehavior
 {
     function let(ExecutionContextInterface $context): void
     {
-        $this->beConstructedWith(['rule_one' => 'rule_one', 'rule_two' => 'rule_two']);
+        $this->beConstructedWith(
+            ['rule_one' => 'rule_one', 'rule_two' => 'rule_two'],
+            ['rule_two' => ['Default', 'rule_two']],
+        );
 
         $this->initialize($context);
     }
@@ -61,5 +66,37 @@ final class PromotionRuleValidatorSpec extends ObjectBehavior
         $constraintViolationBuilder->addViolation()->shouldBeCalled();
 
         $this->validate($promotionRule, new PromotionRule());
+    }
+
+    function it_calls_a_validator_with_group(
+        ExecutionContextInterface $context,
+        PromotionRuleInterface $promotionRule,
+        ValidatorInterface $validator,
+        ContextualValidatorInterface $contextualValidator,
+    ): void {
+        $promotionRule->getType()->willReturn('rule_two');
+
+        $context->getValidator()->willReturn($validator);
+        $validator->inContext($context)->willReturn($contextualValidator);
+
+        $contextualValidator->validate($promotionRule, null, ['Default', 'rule_two'])->willReturn($contextualValidator)->shouldBeCalled();
+
+        $this->validate($promotionRule, new PromotionRule(['groups' => ['Default', 'test_group']]));
+    }
+
+    function it_calls_validator_with_default_groups_if_none_provided_for_promotion_action_type(
+        ExecutionContextInterface $context,
+        PromotionRuleInterface $promotionRule,
+        ValidatorInterface $validator,
+        ContextualValidatorInterface $contextualValidator,
+    ): void {
+        $promotionRule->getType()->willReturn('rule_one');
+
+        $context->getValidator()->willReturn($validator);
+        $validator->inContext($context)->willReturn($contextualValidator);
+
+        $contextualValidator->validate($promotionRule, null, ['Default', 'test_group'])->willReturn($contextualValidator)->shouldBeCalled();
+
+        $this->validate($promotionRule, new PromotionRule(['groups' => ['Default', 'test_group']]));
     }
 }
