+++ tests/Api/Responses/Expected/admin/promotion/post_promotion_with_invalid_rules_response.json
@@ -48,11 +48,6 @@
             "propertyPath": "rules[7].configuration[WEB]",
             "message": "This field is missing.",
             "code": @string@
-        },
-        {
-            "propertyPath": "rules[7].configuration[MOBILE]",
-            "message": "This field was not expected.",
-            "code": @string@
         }
     ]
 }
